section 73

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity'
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -15,12 +16,12 @@ export class TasksService {
   ) {
   }
 
-  public getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto)
+  public getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user)
   }
 
-  async getTaskById(id: string): Promise<Task> {
-    const found = await this.taskRepository.findOne(id)
+  async getTaskById(id: string, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id }})
 
     if (!found) {
       throw new NotFoundException(`Task with id ${id} not found.`)
@@ -29,20 +30,20 @@ export class TasksService {
     return found
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto)
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user)
   }
 
-  async deleteTaskById(id: string): Promise<void> {
-    const deleteResult = await this.taskRepository.delete(id)
+  async deleteTaskById(id: string, user: User): Promise<void> {
+    const deleteResult = await this.taskRepository.delete({ id, userId: user.id })
 
     if (!deleteResult.affected) {
       throw new NotFoundException(`Task with id ${id} not found.`)
     }
   }
 
-  async updateTaskStatus(id: string, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id)
+  async updateTaskStatus(id: string, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user)
     
     task.status = status
